Platforms/Styx: ignore PcdArmPrimaryCoreMask
The PCD PcdArmPrimaryCoreMask is of dubious value, not only because it is
unclear what the point is of masking out bits before comparing a given
MPIDR_EL1 value with a known value belonging to the primary core, but also
because the default value of 0xF03 masks out bits that belong to the actual
affinity levels.

For Styx, simply ignore the mask altogether. Since its version of
ArmPlatformPeiBootAction () records the primary MPIDR_EL1 value upon
first entry, we know its exact value, and no masking is necessary.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Tested-by: Alan Ott <alan@softiron.co.uk>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>