Drivers/Net/MarvellYukon: Add 64-bit DMA support
Add support for 64-bit DMA transfers, since some 64-bit platforms don't
have the ability to generate DMA addresses which can fit in 32-bits.

This code came from the FreeBSD driver, the one from which this driver was
derived.

This patch makes this driver use 64-bit DMA in all cases, because it was
determined that there is no good way to test whether 64-bit DMA is
required.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Alan Ott <alan@softiron.co.uk>
Reviewed-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>