Platforms/ARM/VExpress: make BdsLib dependencies explicit
Before moving to the generic BDS, copy the global BdsLib resolution to
the respective clients so we can drop it later. That way, no new code
will inadvertently bind to it, and it is more obvious which modules
still need to be brought into the 21st century.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Tested-by: Ryan Harkin <ryan.harkin@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>