Platforms/AMD/Styx: fix broken build due to iasl command line change
The upstream BaseTools where modified to allow the use of PCDs in
ASL source code, but this requires the -P option to be removed from
the iasl command line. The Styx platforms need ARCHCC_FLAGS on the
iasl command line, but since we use the single, additive '=', there
is no need to duplicate the other options, and we can safely remove
them.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>