Platforms/FVP-AArch64: switch to simpler DT platform driver
Replace the elaborate but awkward FdtPlatformDxe with the new
DtPlatformDxe, which supports embedded DTBs only, for FVP platforms.
This is done by removing the FdtPlatformDxe reference from
ArmVExpress.dsc.inc and making it a platform-specific decision.

This should be sufficient for virtually all use cases, and if it is
not, there are various way to override the device tree binary presented
to the OS.

As a bonus, this driver makes ACPI and DT mutually exclusive - this
can be configured via the setup screen.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Tested-by: Ryan Harkin <ryan.harkin@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>