Platforms/Marvell: add library class for non-discoverable device init routines
To abstract the initialization required for non-discoverable devices, which
is often platform specific (i.e., enabling VBUS gpios for USB), introduce
a PciEmulationInitLib for use by the PciEmulation code, for which each
platform can supply its own version.

Add dummy version of the library for A70x0 platform in order to satisfy
compilation. It will be extended in future with proper VBUS control,
as well as with the support for other boards.

Change-Id: I7f00f93c4b618f0c0a05ae452fce6983e1a95214
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Marcin Wojtas <mw@semihalf.com>