Application/MdioCmd: add mdio command
Marvell Armada SoC has SMI bus to control the PHY/Switch
that connects to it.
This patch introduces application of mdio command
to configure (read/write) PHY/Switch registers, with
protocol Mdio.
The command takes input Bus number, Address and Register
number, to read or write its value.

There is a dedicate command to read counters for Marvell
soho switch, which is by default disabled, and should only
be enabled for the boards with it.

Change-Id: Ida777de3604a0b04d50dacba8d2b60dca3707d5f
Signed-off-by: jinghua <jinghua@marvell.com>
Contributed-under: TianoCore Contribution Agreement 1.0
Reviewed-on: http://vgitil04.il.marvell.com:8080/42531
Reviewed-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>