Marvell/Armada80x0McBin: Add PP2 description in Dsdt table
This commit introduces ACPI description for PP2 controllers on
Armada-8040-McBin board, which is handled in Linux by mvpp2 driver.
A new Hardware ID for the controller is added: "MRVL0110"

Until the PHY handling for ACPI is developed in Linux, port status
can be handled only with usage of the link interrupts.

Change-Id: Iec2b096b387b45a62d413152b64d65017bceafd9
Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/50076
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>