Marvell/Armada70x0Db: Describe PP2 ports in Dsdt table
This commit introduces ACPI description for PP2 controllers on
Armada-7040-Db board, which is handled in Linux by mvpp2 driver.
A new Hardware ID for the controller is added: "MRVL0110"

Until the PHY handling for ACPI is developed in Linux, port status
can be handled only with usage of the link interrupts. Because
of that, the RGMII port will operate properly only in case
the ethernet cable is plugged during the UEFI stage.

Change-Id: I389802afb6ed786ada3a3a000d8de77d35e0ca41
Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/50077
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>