Marvell/Armada7k8k: Ensure proper access to memory mapped SPI
It turned out that additional configuration of the AMB Bridge
is required when accessing the memory-mapped SPI. The BootROM
however does it by default only when booting from SPI. Otherwise
(e.g. when booting from SD card) proper access to the SPI
data is impossible, which results in ASSERT during FV contents
validation.

Fix above issue during platform initialization. Note, the final
fix will be part of the ARM-TF. Before it's availabe in the
public branch, keep the fix in the UEFI code as well.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Marcin Wojtas <mw@semihalf.com>