Marvell/Drivers: MvGpioDxe: Fix gpio.h solo build failure
There is a inline help routine MarvellGpioGetHandle
in Protocol/gpio.h, to ease selection of the desired
GPIO controller. It makes gpio.h could not build
without including any other header file. Usually it
is OK, since the source file that includes gpio.h
does the header file including. But it is
not convinient since some header files might include
gpio.h only for the basic data structure define, like
GPIO controller type, and it is not nice for it to
do a lot of other including to pass the compilation.
This patch includes necessary header file in gpio.h
to make it pass compilation alone.

Change-Id: If86348b5a8987f69f688ba79a93d05bf80302666
Signed-off-by: jinghua <jinghua@marvell.com>
Contributed-under: TianoCore Contribution Agreement 1.1
Reviewed-on: http://vgitil04.il.marvell.com:8080/52787
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Marcin Wojtas <mw@semihalf.com>