Marvell/Armada7k8k: Use DSB in IO transactions
This patch add dsb after every read to IO, this change
help to solve the ECAM issue. The dsb will make sure
that read transaction reach the target address.
This patch enables using data sync barriers on IO for
the MacchiatoBin board. In order to achieve that,
the ARM-specific part of the generic IoLib was imported
and modified accordingly. The usage is enabled
with MVEBU_PCIE_ECAM_WA build flag.

Change-Id: I7e680359365fa214cbc1c9298a0ba467055a480b
Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Hanna Hawa <hannah@marvell.com>
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/57368
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>