Marvell/Applications: Extend functionality of fupdate to flash on SD
Extend fupdate shell application to be able to flash firmware
on SD and MMC cards. Device on which firmware will be updated
is selected via third argument of command. When no device
is selected spi is flashed.
Command uses BlockIo protocol to communicate with SD/MMC.

Change-Id: I3bf90fa307a8ac45f116b9cf602a391e3e81fb65
Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Tomasz Michalec <tm@semihalf.com>
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/59637
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>