Marvell/Armada7k8k: Enable global configuration of PCIE
This patch enables configuration of the ACPI tables,
as well as the customized PciExpressLib in a way
to make use of global PCDs. Now usage of the ECAM
shift solution can be controlled from a single place
and it is enabled by default. In order to withdraw it
MVEBU_PCIE_ECAM_WA build flag must be used.

Change-Id: Ic298e39245d6fc6222cb2d794c8462d3aa443eb2
Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/57277
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>